@@ -1,11 +1,11 @@
 class UpDownPageLimit {
+  const UpDownPageLimit(this.upLimit, this.downLimit);
   final int upLimit;
   final int downLimit;
-  const UpDownPageLimit(this.upLimit, this.downLimit);
 }
 
 class UpDownButtonEnableState {
+  const UpDownButtonEnableState(this.upState, this.downState);
   final bool upState;
   final bool downState;
-  const UpDownButtonEnableState(this.upState, this.downState);
 }
